update server with structs for jsonapi
